agent-langgraph/requirements.txt
# Shared with agent-cli and agent-mcp (unchanged)
anthropic>=0.84.0
python-dotenv>=0.21.0
pandas>=2.2.2
pyyaml>=6.0.1
fastembed>=0.4.0
asyncpg>=0.29.0

# LangGraph stack — the key differentiator from agent-mcp and agent-cli
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0

# Postgres checkpointer for interrupt() persistence
# Uses psycopg3 (the psycopg package) — coexists with asyncpg (used by store.py)
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.1.0
